requirements.txt
Django
djangorestframework
channels
gunicorn
daphne
boto3
uvicorn[standard]